chore: move LoginForm logic to LoginFormContainer.
