+++ src/components/auth/LoginForm/LoginForm.tsx
@@ -1,10 +1,7 @@
 "use client";
 
 // External deps
-import { SubmitHandler, useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import { useRouter } from "next/navigation";
+import type { UseFormReturn } from "react-hook-form";
 
 // Internal deps
 import { Input } from "@/components/ui/input";
@@ -19,67 +16,23 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { loginService } from "@/api/auth/services";
-import { setToken } from "@/lib/serverActions";
-import { ROUTES } from "@/lib/constants";
 import { Icons } from "@/components/ui/icons";
+import type { LoginFormFields } from "@/containers/auth/LoginContainer/LoginContainer";
 
-const schema = z.object({
-  email: z
-    .string({ required_error: "El email es requerido" })
-    .email({ message: "Ingresa un email valido" }),
-  password: z
-    .string({ required_error: "La contraseña es requerida" })
-    .min(5, { message: "La contraseña debe tener al menos 5 caracteres" }),
-  remember: z.boolean({}).default(false).optional(),
-});
+interface Props {
+  form: UseFormReturn<LoginFormFields>;
+  onSubmit: () => void;
+}
 
-type FormFields = z.infer<typeof schema>;
-
-function LoginForm() {
-  const form = useForm<FormFields>({ resolver: zodResolver(schema) });
+function LoginForm({ form, onSubmit }: Props) {
   const {
     control,
-    handleSubmit,
     formState: { errors, isSubmitting },
-    setError,
   } = form;
-  const router = useRouter();
-
-  const onSubmit: SubmitHandler<FormFields> = async (values) => {
-    try {
-      const response = await loginService({
-        email: values.email,
-        password: values.password,
-      });
-
-      if (!response.success) {
-        throw new Error(response.message);
-      }
-
-      setToken(response.data.token);
-
-      router.push(ROUTES.HOME.path);
-    } catch (error: any) {
-      console.error(error);
-      if (error.message.includes("500")) {
-        setError("root", {
-          message: "Algo salió mal. Inténtalo nuevamente.",
-        });
-      } else {
-        setError("root", {
-          message: "Parece que alguno de tus datos es incorrecto.",
-        });
-      }
-    }
-  };
 
   return (
     <Form {...form}>
-      <form
-        onSubmit={handleSubmit(onSubmit)}
-        className="flex h-full w-full flex-col justify-between"
-      >
+      <form onSubmit={onSubmit} className="flex h-full w-full flex-col justify-between">
         <div className="grid gap-6">
           {/* Email */}
           <FormField
@@ -157,7 +110,7 @@ function LoginForm() {
             className="bg-accent text-accent-foreground hover:bg-accent/90"
             disabled={isSubmitting}
           >
-            {isSubmitting ? <Icons.loader size={24} className="mr-2 animate-spin" /> : "Ingresar"}
+            {isSubmitting ? <Icons.Loader size={24} className="mr-2 animate-spin" /> : "Ingresar"}
           </Button>
         </div>
       </form>
